Tubes.Par

Documentation

broadcast :: (Functor f, Monad m) => Tube a b m r -> f (Tube b c m r) -> f (Tube a c m r)

Broadcast the values of a source Task to Functor of sink Tasks

merge :: (Foldable t, Monad m) => t (Tube a c m s) -> Tube a c m ()

Naively merges the values of some Functor of source Tubes

(*<) :: (Functor f, Monad m) => Tube a b m r -> f (Tube b c m r) -> f (Tube a c m r) infixl 9

Infix synonym for broadcast

(>*) :: (Foldable t, Monad m) => t (Tube a b m s) -> Tube b c m () -> Tube a c m () infixr 0

merges a Functor of source Tubes and pipes (><) to a sink Tube