Tubes.Pump

Documentation

type Pump a b = CofreeT (PumpF a b)

A Pump is the dual to a Tube: where a Tube is a computation manipulating a stream of values, a Pump can be situated on either end of a tube to both insert values when requested and handle any yielded results.

data PumpF a b k

Constructors

PumpF

Fields

recvF :: (a, k)

sendF :: b -> k

class (Functor f, Functor g) => Pairing f g | f -> g, g -> f where

Lovingly stolen from Dan Piponi and David Laing. This defines a poor man's adjunction: it allows adjoint functors to essentially annihilate one another and produce a final value.

If something equivalent comes along, I'll switch to it.

Methods

pair :: (a -> b -> r) -> f a -> g b -> r

pairEffect :: (Pairing f g, Comonad w, Monad m) => (a -> b -> r) -> CofreeT f w a -> FreeT g m b -> m r

pump :: Comonad w => w a -> (w a -> (b, w a)) -> (c -> w a) -> Pump b c w a

recv :: Comonad w => Pump a b w r -> (a, Pump a b w r)

send :: Comonad w => Pump a b w r -> b -> Pump a b w r