Tubes.Source

Documentation

newtype Source m a

An exhaustible source of values parameterized over a base monad. It never awaits, it only yields.

Sources are monad transformers in their own right, as they are possibly finite. They may also be synchronously merged:

    src1 :: Source IO String
    src1 = Source $ each ["line A1", "line A2", "line A3"]

    src2 :: Source IO String
    src2 = Source $ each ["line B1", "line B2", "line B3", "line B4"]

    src3 :: Source IO String
    src3 = src1 merge src2

    main :: IO ()
    main = runTube $ sample src3 >< pour display
    -- line A1
    -- line B1
    -- line A2
    -- line B2
    -- line A3
    -- line B3
    -- line B4

If one source runs out, the other will continue until completion.

Digression: originally merge was the implementation for mappend and '(<>)'. However because Source is ultimately a list transformer I thought it better that these instances preserve the behavior found in lists and instead provide a separate function for synchronous merging.

Constructors

Source

Fields

sample :: Tube () a m ()

reduce :: Monad m => (b -> a -> b) -> b -> Tube () a m () -> m b

Strict left-fold of a Source, using a Pump internally.

merge :: Monad m => Source m a -> Source m a -> Source m a

Interleave the values of two Sources until both are exhausted.