Tubes.Util

Documentation

cat :: Monad m => Tube a a m r

Continuously relays any values it receives. Iteratee identity.

map :: Monad m => (a -> b) -> Tube a b m r

Transforms all incoming values according to some function.

drop :: Monad m => Int -> Tube a a m r

Refuses to yield the first n values it receives.

take :: Monad m => Int -> Tube a a m ()

Relay only the first n elements of a stream.

takeWhile :: Monad m => (a -> Bool) -> Tube a a m ()

Terminates the stream upon receiving a value violating the predicate

filter :: Monad m => (a -> Bool) -> Tube a a m r

Yields only values satisfying some predicate.

reduce

Arguments

:: Monad m

=> (x -> a -> x)
step function

-> x
initial value

-> (x -> b)
final transformation

-> Source a m ()
stream source

-> m b

Strict left-fold of a stream. Note that the actual return type of the source is not relevant, only the intermediate yield type.

every :: (Foldable t, Monad m) => t b -> Tube a (Maybe b) m ()

Similar to each except it explicitly marks the stream as exhausted

prompt :: Source String IO ()

Source of Strings from stdin. This is mostly for debugging / ghci example purposes.

display :: Sink String IO ()

Sink for Strings to stdout. This is mostly for debugging / ghci example purposes.

unyield :: Monad m => FreeT (TubeF x b) m () -> m (Maybe (b, FreeT (TubeF x b) m ()))

Taps the next value from a source, maybe.